+++ docs/source/src_mock_py/microphysics_scheme_wrapper.rst
@@ -1,7 +1,7 @@
 Wrapper for Python Mock Microphysics Scheme
 ============================================
 
-.. autoclass:: libs.src_mock_py.microphysics_scheme_wrapper.MicrophysicsSchemeWrapper
+.. autoclass:: libs.mock_microphys.microphysics_scheme_wrapper.MicrophysicsSchemeWrapper
    :members:
    :undoc-members:
    :private-members:
