+++ docs/source/src_mock_py/mock_microphysics_scheme.rst
@@ -1,7 +1,7 @@
 Python Mock Microphysics Module
 ===============================
 
-.. autoclass:: libs.src_mock_py.mock_microphysics_scheme.MicrophysicsScheme
+.. autoclass:: libs.mock_microphys.mock_microphysics_scheme.MicrophysicsScheme
    :members:
    :undoc-members:
    :private-members:
