+++ libs/mock_microphys/__init__.py
@@ -3,7 +3,7 @@ Copyright (c) 2024 MPI-M, Clara Bayley
 
 ----- Microphysics Test Cases -----
 File: __init__.py
-Project: src_mock_py
+Project: mock_microphys
 Created Date: Tuesday 27th February 2024
 Author: Clara Bayley (CB)
 Additional Contributors:
