+++ libs/mock_microphys/mock_microphysics_scheme.py
@@ -3,7 +3,7 @@ Copyright (c) 2024 MPI-M, Clara Bayley
 
 ----- Microphysics Test Cases -----
 File: mock_microphysics_scheme.py
-Project: src_mock_py
+Project: mock_microphys
 Created Date: Tuesday 27th February 2024
 Author: Clara Bayley (CB)
 Additional Contributors:
